Create HW 3
Assignment 3, due on 10/14/2020 at 11:59 pm
Instructions:
Write the commands needed to solve the following problems on mis01.scu.edu. Before solving each
exercise, position yourself in /home/OMIS107/Lecture2.
• All of these exercises are on alice.txt.
• Unless otherwise specified, don’t worry about uppercase/lowercase.
Problem 1
Position yourself in /home/OMIS107/Lecture2. Print the lines of alice.txt with at least two words
starting with “a”, with the first word being composed of at least 8 more letters after “a”.
Problem 2
Add “his or her majesty” before the occurrences of the word “the” + word starting with a capital letter
and the word “a” + word starting with a capital letter, as in these examples:
• “the Hatter” → “his or her majesty the Hatter”
• “a Caterpillar” → “his or her majesty a Caterpillar”
• “Vanilla Ice cream” → “Vanilla Ice cream” (unchanged because “a” is not a word)
Problem 3
To emphasize a verb in past tense, one could add “did” before it and turn the verb into present tense.
Assume that all words ending in “ed” are verbs in past tense (e.g., invented, asked). Find all words
ending in “ed”, add “did” before them and turn them into present tense. For example:
• “You invented it” → “you did invent it”
• “Alice asked a question” → “Alice did ask a question”
• “She asked.” → “She did ask.”
• “They both cried.” → “They both did cri.” (do not worry about whether removing “ed” is
grammatically incorrect)
Problem 4
The goal of this exercise is to replace the occurrences in alice.txt that contain the word “said” + other
words + punctuation mark with ‘said someone’ + punctuation mark. By “punctuation mark”, we mean
any of the following characters ,.:;
Some examples:
1. Line 3300: 'I won't!' said Alice. → 'I won't!' said someone.
2. Line 3314: 'Wake up, Alice dear!' said her sister; 'Why, what a long sleep you've →
 'Wake up, Alice dear!' said someone; 'Why, what a long sleep you've
3. Line 396: but it said nothing. → but it said someone.
If the word “said” is followed by more than one punctuation mark in the same line, make sure to replace
only up to the first punctuation mark. That is, the following replacement for example 2 is wrong:
 'Wake up, Alice dear!' said someone, what a long sleep you've