Create HW 5
Assignment 5, due on 10/27/2020 at 10:20 am
Instructions:
Write the commands needed to solve the following exercises on mis01.scu.edu.
Problem 1
Find the number of existing processes. Your output should be formatted as follows:
On xxx, there were yyy processes
xxx is the current timestamp (which you can retrieve with the command date)
yyy is the number of processes
For example: On Thu Oct 22 12:15:52 PDT 2020, there were 271 processes
Problem 2
The file /home/OMIS107/HW5/name contains one line of text. How much RAM is used by the processes (all
processes) whose names contain that text? Your code should return just a number from 0 to 100. For instance, if
/home/OMIS107/HW5/name contains the text ps, then your code should return the total % of RAM (0-100) used by
all processes whose name contains ps. Your code must work correctly even if the content of
/home/OMIS107/HW5/name is changed.
Problem 3 (1 point)
Position yourself in /home. Print only the file/directory name, followed by a colon, followed by the “group” permissions
(the middle three permission bits) for all of the files/directory owned by the current user (whomever the user may be).
Example: the output of the command executed by msamorani is this:
Problem 4 (1 point)
Read file hw5 inside your directory. Create one problem (text + solution) as indicated in that file and then upload your
problem and solution on this shared document.
The goal is to help fellow students to prepare for the midterm; so, try to make the problem useful. The text of your
problem should be as clear as possible and as short as possible. Avoid creating problems that are too trivial or too
difficult; the difficulty of your problem should be similar to the problems in this and past assignments.
My intention is to give everyone full marks; the only way not to get full marks are the following:
• You do not perform this task
• You do not include a solution
• Your problem is clearly too trivial. For example, “print the current date” (solution: date), “print the list of
existing processes” (solution: ps -efl).
If the solution you provide is incorrect, you will still receive full marks. But you must make an honest effort to get it right.