Create HW 6
Assignment 6, due on 11/11/2020 at 11:59 pm
Instructions:
Write the script needed to solve the following problem on mis01.scu.edu.
Scripting Problem (4 points)
The goal of this problem is to find the current user’s group with the largest number of members. That is,
considering the groups that the current user belongs to (with the exclusion of the group users), which one is
composed of the largest number of members? (Don’t worry about ties)
Start by looking at the file /etc/group, which contains the members belonging to each in the following
format: groupname:x:groupid:comma_separated_list_of_members
For example, the entry corresponding to group OMIS107_ADMIN is as follows:
OMIS107_ADMIN:x:1003:msamorani,scu\msamorani,ywang6,labinstructor,...
Your output should be formatted as follows:
The current user's most numerous group is xxxx, which has yyyy members
Note: don’t worry about explicitly excluding the group users, because that group is not present in
/etc/group.